Fix few issues with network id
