+++ src/views/Stake/stake.scss
@@ -72,11 +72,13 @@
       padding-top: 0;
       padding-bottom: 0;
       position: absolute;
-      left: 105px;
-      // transition: left 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms, transform 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-      // &.Mui-expanded {
-      //   left: -32px
-      // }
+      left: 100px;
+      top: -3px;
+      transition: left 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms, top 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms, transform 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
+      &.Mui-expanded {
+        left: 104px;
+        top: 1px;
+      }
     }
 
     .MuiAccordionDetails-root {
